Assignment2/requirements.txt
# Python Dependencies for CVI620 Assignment 2

# Core ML/DL frameworks
tensorflow>=2.10.0
numpy>=1.23.0
pandas>=1.5.0

# Computer Vision
Pillow>=9.2.0
opencv-python>=4.6.0

# Machine Learning
scikit-learn>=1.1.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Progress bars (optional)
tqdm>=4.64.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipykernel>=6.15.0
